added firebase to server, login page to flutter
