+++ Flutter/biteme/android/app/src/main/AndroidManifest.xml
